Version 7.1 update

+++ NeuralNetwork_PERF.h
@@ -33,7 +33,7 @@ private:
 	double m_weights[GLOBAL_WC];
 	double m_king_val;
 	int m_weight_iter;
-	int m_player_modifier;
+	double m_player_modifier;
 };
 
 #endif
